compsecassignment/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.59.0
django-environ==0.11.2
django-recaptcha==4.0.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
psycopg==3.1.16
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
